[Fix] 스테이지 매니저에 튜토리얼 지역에 몬스터&FO들이 갱신되지 않던 문제 수정
- 튜토리얼지역 StageData에 stageKey가 LivingRoom으로 되어있어서 발생한 문제였음.

+++ Assets/Scripts/FarmingSystem/Interactable_Farming.cs
@@ -16,8 +16,9 @@ public class Interactable_Farming : InteractableBase, ISpawnable
 
     public Action DeactiveAction { get; set; }
     public Transform OriginTransform { get; set; }
+    bool isUsed;
 
-        
+
 
     // 드랍테이블 정보에 기입된 개수만큼 아이템 인스턴스 생성 반복
     public void DropItemInstances(DropInfo dropInfo)
@@ -61,6 +62,9 @@ public class Interactable_Farming : InteractableBase, ISpawnable
     public override void Interact()
     {
         base.Interact();
+
+        if (isUsed) return;
+        isUsed = true;
         DropInfo dropInfo = dropTable.GetDropItemInfo();
 
         switch (dropType)
